==> 1 - Creation of Single Data Set.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline

# Importing the Data 

data_13 = pd.read_excel("../Data/MATERNITY_2013.xlsx",index_col="admit_year")
data_14 = pd.read_excel("../Data/MATERNITY_2014.xlsx",index_col="admit_year")
data_15 = pd.read_excel("../Data/MATERNITY_2015.xlsx",index_col="admit_year")
data_16 = pd.read_excel("../Data/MATERNITY_2016.xlsx",index_col="admit_year")
data_17 = pd.read_excel("../Data/MATERNITY_2017.xlsx",index_col="admit_year")
data_18 = pd.read_excel("../Data/MATERNITY_2018.xlsx",index_col="admit_year")
data_19 = pd.read_excel("../Data/MATERNITY_2019.xlsx",index_col="admit_year")
data_20 = pd.read_excel("../Data/MATERNITY_2020(jan-mar).xlsx",index_col="admit_year")


scheme_B=pd.read_excel("../Data/B_MATERNITY_CLAIMS.xlsx",index_col="admit_year")
scheme_C=pd.read_excel("../Data/C_MATERNITY_MEM_CLAIMS.xlsx",index_col="admit_year")

# Re-Formatting the Column Headings

data_13.columns=data_13.columns.str.title()
data_14.columns=data_14.columns.str.title()
data_15.columns=data_15.columns.str.title()
data_16.columns=data_16.columns.str.title()
data_17.columns=data_17.columns.str.title()
data_18.columns=data_18.columns.str.title()
data_19.columns=data_19.columns.str.title()
data_20.columns=data_20.columns.str.title()

scheme_B.columns=scheme_B.columns.str.title()
scheme_C.columns=scheme_C.columns.str.title()

data_13.drop_duplicates()
data_14.drop_duplicates()
data_15.drop_duplicates()
data_16.drop_duplicates()
data_17.drop_duplicates()
data_18.drop_duplicates()
data_19.drop_duplicates()
data_20.drop_duplicates()

scheme_B.drop_duplicates()
scheme_C.drop_duplicates()

### Preview of the 2019 Data Frame

data_19.head(3)

# Joining All Data Together

Joining All Scheme A's Data Together

scheme_A=pd.concat(objs=[data_13,data_14,data_15,data_16,data_17,data_18,data_19])

scheme_A.head(3)

Joining Scheme A, B and C Together

all_schemes=pd.concat(objs=[scheme_A,scheme_B,scheme_C],keys=["Scheme_A", "Scheme_B", "Scheme_C"])

all_schemes

# Exporting This Single Data Frame as an Excel Spreadsheet 

excel_file = pd.ExcelWriter("All_Schemes_Data.xlsx")

all_schemes.to_excel(excel_file,sheet_name="Sheet1")

excel_file.save()
