==> 2 - Cleaning Process.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline

# Importing the Required Data 

data=pd.read_excel("../Data/All_Schemes_Data.xlsx",index_col=[0,1])
data.head(3)

# Cleaning the Data

### Formatting the Headings for Consistency

data.columns=data.columns.str.title()

### Removing Any Fully Duplicated Rows 

data.drop_duplicates()

# Note: there were actually no fully duplicated rows

### Looking at the Form of the Data Columns and Identifying the Number of Null Values in Each Column

data.info()

### Re-defining the Type of Values Under Each Column to Simply Calculations and Reduce Size of Dataset

data.columns

data["Admission_Category"] = data["Admission_Category"].astype("category")
data["Hosp_Group"] = data["Hosp_Group"].astype("category")
data["Magisterial_District_Descr"] = data["Magisterial_District_Descr"].astype("category")
data["Region_Descr"] = data["Region_Descr"].astype("category")
data["Final_Diagnosis"] = data["Final_Diagnosis"].astype("category")
data["Patient_Gender"] = data["Patient_Gender"].astype("category")

# Creating a New Column That Combines Member Key and Beneficiary Number to Give a Unique Member ID

data["Member_ID"] = data["Member_Key"].astype(str) +"-"+ data["Beneficiary_Number"].astype(str)
data

# Removing Data Points That Are Not Within South Africa

data["Region_Descr"].unique().tolist()

data.replace(to_replace="KWAZULU-NATAL",value="KWAZULU NATAL",inplace=True)
data.replace(to_replace="NORTHWEST PROVINCE",value="NORTH WEST",inplace=True)

data["Region_Descr"].unique().tolist()

req1=data["Region_Descr"]=="NORTHERN CAPE"
req2=data["Region_Descr"]=="KWAZULU NATAL"
req3=data["Region_Descr"]=="EASTERN CAPE"
req4=data["Region_Descr"]=="GAUTENG"
req5=data["Region_Descr"]=="LIMPOPO"
req6=data["Region_Descr"]=="FREE STATE"
req7=data["Region_Descr"]=="NORTH WEST"
req8=data["Region_Descr"]=="MPUMALANGA"
req9=data["Region_Descr"]=="WESTERN CAPE"

data=data[req1|req2|req3|req4|req5|req6|req7|req8|req9]

data["Region_Descr"].unique()

## Splitting the Data According to Mother and Child

mom_data=data[data["Age_At_Treatment"]>10]

baby_data=data[data["Age_At_Treatment"]<=10]

baby_data

# Removing Problematic Rows

Remove any rows where the admission category for mother contain newborn

mom_data["Admission_Category"].unique()

mom_data[mom_data["Admission_Category"]=="Newborn Care"]

mom_data[mom_data["Admission_Category"]=="Newborn - Complications/Premature baby minor"]

mom_data[mom_data["Admission_Category"]=="Complex Newborn Care"]

mom_data[mom_data["Admission_Category"]=="Newborn - Uncomplicated Birth"]

mom_data[mom_data["Admission_Category"]=="Newborn - Complications/Premature baby major"]["Final_Diagnosis"]

Removing Rows in mom_data Whose Admission Category are Related to Newborn Care 

cond1=mom_data["Admission_Category"]!="Newborn Care"
cond2=mom_data["Admission_Category"]!="Newborn - Uncomplicated Birth"
cond3=mom_data["Admission_Category"]!="Newborn - Complications/Premature baby minor"
cond4=mom_data["Admission_Category"]!="Newborn - Complications/Premature baby major"
cond5=mom_data["Admission_Category"]!="Complex Newborn Care"

mom_data=mom_data[cond1 & cond2 & cond3 & cond4 & cond5]

mom_data["Admission_Category"].unique()

19/len(mom_data)

Removing Rows in baby_data, Where the Admission Category is Complictaions of the Pueperium

baby_data["Admission_Category"].unique()

baby_data=baby_data[baby_data["Admission_Category"]!="Complications of the pueperium"]

# Saving the Mom Data and Baby Data into 2 Seperate Excel Files

excel_file = pd.ExcelWriter("All_Schemes_Data_Broken_Into_Mom_And_Baby.xlsx")

mom_data.to_excel(excel_file,sheet_name="Mom_Data")
baby_data.to_excel(excel_file,sheet_name="Baby_Data")

excel_file.save()
