==> 4 - Analysis of Infants.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline

mom_data=pd.read_excel("../Data/All_Schemes_Data_Broken_Into_Mom_And_Baby.xlsx", sheet_name="Mom_Data",index_col=[0,1])
baby_data=pd.read_excel("../Data/All_Schemes_Data_Broken_Into_Mom_And_Baby.xlsx", sheet_name="Baby_Data",index_col=[0,1])

baby_data.head(3)

# Analysing the Newborn Complications

print(f'Number of infants charged for complications: {baby_data["Member_ID"].nunique()}')

baby_data["Final_Diagnosis"].value_counts()

## Complications Caused During Labour and Delivery 

print(f'Number of infants affected by caesarean section: {len(baby_data[baby_data["Final_Diagnosis"]=="P03.4"])}')

print(f'Number of infants affected by forceps delivery: {len(baby_data[baby_data["Final_Diagnosis"]=="P03.2"])}')

print(f'Number of infants affected by vacuum extracter: {len(baby_data[baby_data["Final_Diagnosis"]=="P03.3"])}')

print(f'Number of infants affected by breech delivery and extraction: {len(baby_data[baby_data["Final_Diagnosis"]=="P03.0"])}')

print(f'Number of infants affected by other malpresentation, malposition and disproportion during labour and delivery: {len(baby_data[baby_data["Final_Diagnosis"]=="P03.1"])}')

## Rispiratory Distress Syndrome of Newborn P22

RDS=pd.concat([baby_data[baby_data["Final_Diagnosis"]=="P22.0"],baby_data[baby_data["Final_Diagnosis"]=="P22.1"],baby_data[baby_data["Final_Diagnosis"]=="P22.8"],baby_data[baby_data["Final_Diagnosis"]=="P22.9"]])


print(f' Average cost of Respitory Distress Syndrome: {round(RDS["Total_Cost"].mean(),2)}')

print(f'Number of infants that experienced respiratory distress syndrome: {len(RDS)}')

print(f'Percentage of infants in the data set that experienced respiratory distress syndrome: {round(len(RDS)/baby_data["Member_ID"].nunique(),4)}')

RDS_1=RDS[["Member_Key","Admit_Month"]]
mom_data_1=mom_data[["Member_Key","Admit_Month","Final_Diagnosis","Admission_Category"]]

merged=RDS_1.merge(mom_data_1,how="inner",on=["Member_Key","Admit_Month"])
merged["Final_Diagnosis"].value_counts() 
# This is the final diagnosis of the mom

merged["Member_Key"].nunique()

print(f'Number of infants that experienced respiratory distress syndrome who were born via caeserean section: {len(merged[merged["Admission_Category"]=="Caesarean Delivery"])}')

print(f'Number of babys that experienced respiratory distress syndrome who were born through a natural delivery: {len(merged[merged["Admission_Category"]=="Normal Delivery"])}')

This confrims what was said in literature review about caeserian section babies experiencing significantly worse lung function

print(f'Number of babys that experienced respiratory distress syndrome whose mothers experienced maternal care due to uterine scar from previous surgery:{828}')


print(f'Number of babys that experienced respiratory distress syndrome whose mothers experienced pre-term labour without delivery:{369}')

## Neonatal Jaundice P59.9

This is not further looked into because there is no association between neonatal Jaundice and the Method of Delivery.

## Preterm Infants P07.3

This is not further looked into because there is the method of delivery does not affect whether a baby is born pre-term

## HIV related

HIV_baby=baby_data[baby_data["Member_Key"].isin(mom_data[mom_data["Final_Diagnosis"]=="O98.7"]["Member_Key"])]
HIV_baby
# This is all the babies whose moms were recorded as havign HIV during pregnancy or delivery 

print(f'Average cost to infants whose mother was recorded as HIV positive: {round(HIV_baby["Total_Cost"].mean(),2)}')

HIV_baby["Final_Diagnosis"].value_counts()

9 of the babies suffered from bacterial sepsis - likley from mom having HIV

8 of the babies were born pre-term

# Feeding Problems

 {
    'P92.0':len(baby_data[baby_data["Final_Diagnosis"]=="P92.0"]),
    'P92.1':len(baby_data[baby_data["Final_Diagnosis"]=="P92.1"]),
    'P92.2':len(baby_data[baby_data["Final_Diagnosis"]=="P92.2"]),
    'P92.3':len(baby_data[baby_data["Final_Diagnosis"]=="P92.3"]),
    'P92.4':len(baby_data[baby_data["Final_Diagnosis"]=="P92.4"]),
    'P92.5':len(baby_data[baby_data["Final_Diagnosis"]=="P92.5"]),
    'P92.8':len(baby_data[baby_data["Final_Diagnosis"]=="P92.8"]),
    'P92.9':len(baby_data[baby_data["Final_Diagnosis"]=="P92.9"])
}

feeding=pd.concat([baby_data[baby_data["Final_Diagnosis"]=="P92.0"],baby_data[baby_data["Final_Diagnosis"]=="P92.1"],baby_data[baby_data["Final_Diagnosis"]=="P92.2"],baby_data[baby_data["Final_Diagnosis"]=="P92.3"],baby_data[baby_data["Final_Diagnosis"]=="P92.4"],baby_data[baby_data["Final_Diagnosis"]=="P92.5"],baby_data[baby_data["Final_Diagnosis"]=="P92.8"],baby_data[baby_data["Final_Diagnosis"]=="P92.9"]])

feeding1=mom_data[mom_data['Member_Key'].isin(feeding['Member_Key'])]

print(f'Number of infants who experienced feeding problems: {feeding["Member_ID"].nunique()}')

feeding1["Admission_Category"].value_counts()

print(f'Average cost of feeding problems: {round(feeding["Total_Cost"].mean(),2)}')
