==> 5 - Creation of Data Sets for Predictions.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline

from scipy.stats import zscore

data=pd.read_excel("../Data/All_Schemes_Data_Broken_Into_Mom_And_Baby.xlsx", sheet_name="Mom_Data",index_col=[0,1])

# Removing Outlires


adm_cat=data.groupby("Admission_Category")
Natural=adm_cat.get_group("Normal Delivery")
Caesarean=adm_cat.get_group("Caesarean Delivery")

total_cost_df_Normal=Natural["Total_Cost"]
total_cost_df_Caesarean=Caesarean["Total_Cost"]

z_scores = zscore(total_cost_df_Normal)
abs_z_scores = np.abs(z_scores)
filtered_entries = (abs_z_scores < 3)
Natural = Natural[filtered_entries]
Natural

z_scores = zscore(total_cost_df_Caesarean)
abs_z_scores = np.abs(z_scores)
filtered_entries = (abs_z_scores < 3)
Caesarean = Caesarean[filtered_entries]
Caesarean

Natural=Natural.reset_index()
Natural[["Admit_Month","Total_Cost"]]


Caesarean=Caesarean.reset_index()
Caesarean[["Admit_Month","Total_Cost"]]

excel_file = pd.ExcelWriter("ML_data_for_thesis.xlsx")

Caesarean.to_excel(excel_file,sheet_name="Caesarean Deliveries")
Natural.to_excel(excel_file,sheet_name="Natural")

excel_file.save()
