==> 7 - Analysis of Exposure Data.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline

# Reading in the Data

data_13 = pd.read_excel("MATERNITY_MEM_2013.xlsx")
data_14 = pd.read_excel("MATERNITY_MEM_2014.xlsx")
data_15 = pd.read_excel("MATERNITY_MEM_2015.xlsx")
data_16 = pd.read_excel("MATERNITY_MEM_2016.xlsx")
data_17 = pd.read_excel("MATERNITY_MEM_2017.xlsx")
data_18 = pd.read_excel("MATERNITY_MEM_2018.xlsx")
data_19 = pd.read_excel("MATERNITY_MEM_2019.xlsx")

scheme_B=pd.read_excel("B_MATERNITY_EXPOSURE.xlsx")
scheme_C=pd.read_excel("C_MATERNITY_MEM_EXPOSURE.xlsx")

# Merging the Data

scheme_A_no_13_or_14=pd.concat(objs=[data_15,data_16,data_17,data_18,data_19])

scheme_A_no_13_or_14.head(3)

# Cleaning the Headings

data_13.columns=data_13.columns.str.title()
data_14.columns=data_14.columns.str.title()
scheme_A_no_13_or_14.columns=scheme_A_no_13_or_14.columns.str.title()
scheme_B.columns=scheme_B.columns.str.title()
scheme_C.columns=scheme_C.columns.str.title()

# Removing the Rows Where the Region is Outside of South Africa

scheme_A_no_13_or_14.replace(to_replace="KWAZULU-NATAL",value="KWAZULU NATAL",inplace=True)
scheme_A_no_13_or_14.replace(to_replace="NORTHWEST PROVINCE",value="NORTH WEST",inplace=True)

scheme_A_no_13_or_14["Region_Descr"].unique().tolist()

req1=scheme_A_no_13_or_14["Region_Descr"]=="NORTHERN CAPE"
req2=scheme_A_no_13_or_14["Region_Descr"]=="KWAZULU NATAL"
req3=scheme_A_no_13_or_14["Region_Descr"]=="EASTERN CAPE"
req4=scheme_A_no_13_or_14["Region_Descr"]=="GAUTENG"
req5=scheme_A_no_13_or_14["Region_Descr"]=="LIMPOPO"
req6=scheme_A_no_13_or_14["Region_Descr"]=="FREE STATE"
req7=scheme_A_no_13_or_14["Region_Descr"]=="NORTH WEST"
req8=scheme_A_no_13_or_14["Region_Descr"]=="MPUMALANGA"
req9=scheme_A_no_13_or_14["Region_Descr"]=="WESTERN CAPE"
req10=scheme_A_no_13_or_14["Region_Descr"]=="nan"

scheme_A_no_13_or_14=scheme_A_no_13_or_14[req1|req2|req3|req4|req5|req6|req7|req8|req9|req10]

scheme_A_no_13_or_14["Region_Descr"].unique().tolist()

scheme_A=pd.concat([scheme_A_no_13_or_14,data_13,data_14])

scheme_A

# Breaking Up the Data Into Years

scheme_A["Year"].unique()

scheme_A_13=scheme_A[scheme_A["Year"]==2013]
scheme_A_14=scheme_A[scheme_A["Year"]==2014]
scheme_A_15=scheme_A[scheme_A["Year"]==2015]
scheme_A_16=scheme_A[scheme_A["Year"]==2016]
scheme_A_17=scheme_A[scheme_A["Year"]==2017]
scheme_A_18=scheme_A[scheme_A["Year"]==2018]
scheme_A_19=scheme_A[scheme_A["Year"]==2019]

scheme_B["Year"].unique()

scheme_B_13=scheme_B[scheme_B["Year"]==2013]
scheme_B_14=scheme_B[scheme_B["Year"]==2014]
scheme_B_15=scheme_B[scheme_B["Year"]==2015]
scheme_B_16=scheme_B[scheme_B["Year"]==2016]
scheme_B_17=scheme_B[scheme_B["Year"]==2017]
scheme_B_18=scheme_B[scheme_B["Year"]==2018]
scheme_B_19=scheme_B[scheme_B["Year"]==2019]
scheme_B_20=scheme_B[scheme_B["Year"]==2020]

scheme_C["Year"].unique()

scheme_C_13=scheme_C[scheme_C["Year"]==2013]
scheme_C_14=scheme_C[scheme_C["Year"]==2014]
scheme_C_15=scheme_C[scheme_C["Year"]==2015]
scheme_C_16=scheme_C[scheme_C["Year"]==2016]
scheme_C_17=scheme_C[scheme_C["Year"]==2017]
scheme_C_18=scheme_C[scheme_C["Year"]==2018]
scheme_C_19=scheme_C[scheme_C["Year"]==2019]
scheme_C_20=scheme_C[scheme_C["Year"]==2020]

# Calculating Number of Female Members of Childbearing Age per Year for Each Scheme

Scheme A

{
    'Number of Female Members in 2013':round((scheme_A_13["Exposure"].sum())/12,2),
    'Number of Female Members in 2014':round((scheme_A_14["Exposure"].sum())/12,2),
    'Number of Female Members in 2015':round((scheme_A_15["Exposure"].sum())/12,2),
    'Number of Female Members in 2016':round((scheme_A_16["Exposure"].sum())/12,2),
    'Number of Female Members in 2017':round((scheme_A_17["Exposure"].sum())/12,2),
    'Number of Female Members in 2018':round((scheme_A_18["Exposure"].sum())/12,2),
    'Number of Female Members in 2019':round((scheme_A_19["Exposure"].sum())/12,2),
}

print(f'Average Number of Members per Year : {round(((scheme_A_13["Exposure"].sum())/12+(scheme_A_14["Exposure"].sum())/12+(scheme_A_15["Exposure"].sum())/12+(scheme_A_16["Exposure"].sum())/12+(scheme_A_17["Exposure"].sum())/12+(scheme_A_18["Exposure"].sum())/12+(scheme_A_19["Exposure"].sum())/12)/7,2)}')

Scheme B

{
    'Number of Female Members in 2013':round((scheme_B_13["Exposure"].sum())/12,2),
    'Number of Female Members in 2014':round((scheme_B_14["Exposure"].sum())/12,2),
    'Number of Female Members in 2015':round((scheme_B_15["Exposure"].sum())/12,2),
    'Number of Female Members in 2016':round((scheme_B_16["Exposure"].sum())/12,2),
    'Number of Female Members in 2017':round((scheme_B_17["Exposure"].sum())/12,2),
    'Number of Female Members in 2018':round((scheme_B_18["Exposure"].sum())/12,2),
    'Number of Female Members in 2019':round((scheme_B_19["Exposure"].sum())/12,2),
}

print(f'Average Number of Members per Year : {round(((scheme_B_13["Exposure"].sum())/12+(scheme_B_14["Exposure"].sum())/12+(scheme_B_15["Exposure"].sum())/12+(scheme_B_16["Exposure"].sum())/12+(scheme_B_17["Exposure"].sum())/12+(scheme_B_18["Exposure"].sum())/12+(scheme_B_19["Exposure"].sum())/12)/7,2)}')

Scheme C

{
    'Number of Female Members in 2013':round((scheme_C_13["Exposure"].sum())/12,2),
    'Number of Female Members in 2014':round((scheme_C_14["Exposure"].sum())/12,2),
    'Number of Female Members in 2015':round((scheme_C_15["Exposure"].sum())/12,2),
    'Number of Female Members in 2016':round((scheme_C_16["Exposure"].sum())/12,2),
    'Number of Female Members in 2017':round((scheme_C_17["Exposure"].sum())/12,2),
    'Number of Female Members in 2018':round((scheme_C_18["Exposure"].sum())/12,2),
    'Number of Female Members in 2019':round((scheme_C_19["Exposure"].sum())/12,2),
}

print(f'Average Number of Members per Year : {round(((scheme_C_13["Exposure"].sum())/12+(scheme_C_14["Exposure"].sum())/12+(scheme_C_15["Exposure"].sum())/12+(scheme_C_16["Exposure"].sum())/12+(scheme_C_17["Exposure"].sum())/12+(scheme_C_18["Exposure"].sum())/12+(scheme_C_19["Exposure"].sum())/12)/7,2)}')

# Calculating Number of Members per Year for All Schemes

data=pd.concat(objs=[scheme_A,scheme_B,scheme_C])

data_13=data[data["Year"]==2013]
data_14=data[data["Year"]==2014]
data_15=data[data["Year"]==2015]
data_16=data[data["Year"]==2016]
data_17=data[data["Year"]==2017]
data_18=data[data["Year"]==2018]
data_19=data[data["Year"]==2019]

{
    'Number of Female Members in 2013':round((data_13["Exposure"].sum())/12,2),
    'Number of Female Members in 2014':round((data_14["Exposure"].sum())/12,2),
    'Number of Female Members in 2015':round((data_15["Exposure"].sum())/12,2),
    'Number of Female Members in 2016':round((data_16["Exposure"].sum())/12,2),
    'Number of Female Members in 2017':round((data_17["Exposure"].sum())/12,2),
    'Number of Female Members in 2018':round((data_18["Exposure"].sum())/12,2),
    'Number of Female Members in 2019':round((data_19["Exposure"].sum())/12,2),
}

print(f'Average Number of Members per Year : {round(((data_13["Exposure"].sum())/12+(data_14["Exposure"].sum())/12+(data_15["Exposure"].sum())/12+(data_16["Exposure"].sum())/12+(data_17["Exposure"].sum())/12+(data_18["Exposure"].sum())/12+(data_19["Exposure"].sum())/12)/7,2)}')
